power:qpnp-smbcharger-mmi: treat HVDCP2 as HVDCP3 Sometimes
Porting the changes of CR: (CR) to 3.18 kernel

If a External Battery is present and it supports Parallel
Path then HVDCP detection does not fully function.  HVDCp2 is
the best that can be detected.  So for this small subset
allow the 5V ICL to be set to 3A or treat it as a HVDCP3 would
be.

Change-Id: Iacb36b3153682d63ff87802327e6be87Defendere382b
Signed-off-by: kumarv <kumarv@motorola.com>
Reviewed-on: https://gerrit.mot.com/1308299
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Srihari Sathyanarayana <sriharis@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key